Add a clients module
